Ship battle (#220)
* Partial commit, need to go back to work on action initiation decisions

* Added ShipBattleTargetDecision class

* Added starter for test class

* Added some more functionality for ship battle testing

* Limited rules-based battle action to execute orders phase only

* Implemented ship battle action through damage calculation

* Removed NIU ShipBattleAction class

* Implemented ship battle rotation damage

* Partial implementation of extending facility shields

* Updated ShieldsEvaluator to use float values

* Cleanup of MultiplyEvaluator class

* Added shield extension modifier to ST1ERuleSet

* Fixed bug in PhysicalShipCard.isDockedAt method

* Fixed bug in shield extension implementation

* Added WIP modifier to disable weapons when ship doesn't meet requirements

* Implemented weapons disabling

* Added test for ship with no weapons trying to initiate battle. Fixed bug.

* Fleshing out ship battle rules

+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/cards/physicalcard/PhysicalCard.java
@@ -169,7 +169,7 @@ public interface PhysicalCard extends Filterable {
 
     @JsonIgnore
     default Integer getStrength(DefaultGame cardGame) {
-        return cardGame.getGameState().getModifiersQuerying().getStrength(this);
+        return (int) cardGame.getGameState().getModifiersQuerying().getStrength(this);
     }
 
     void reveal();
